refactor(prompt-management): split PromptEditor into smaller components
Split monolithic PromptEditor.tsx (460 lines) into focused components:

- PromptEditor.tsx: Thin container (~130 lines) for composition
- prompt-editor/hooks/usePromptEditorState.ts: State and mutations
- prompt-editor/PromptEditorForm.tsx: Form inputs and validation UI
- prompt-editor/PromptEditorActions.tsx: Save, Test, Bulk Evaluate buttons
- prompt-editor/PromptVersionInfo.tsx: Langfuse version metadata
- prompt-editor/BulkEvaluationSection.tsx: Job progress display

This improves maintainability, testability, and follows single
responsibility principle.
